refactor(portal-frontend): hoist GitHubConfig + WebCrawlerConfig to -connector-types
Follow-up to SPEC-PORTAL-CONNECTOR-WIZARD-EXTRACT-001 (F-C2 partial).

Both interfaces are wizard-only (used by add-connector + edit-connector,
zero KB-tab consumers). Per portal-frontend.md § "File organization for
shared types and helpers" — smallest-shared-scope is the parent route
directory, not $kbSlug/-kb-types.ts (which is for KB-tab routes).

The original SPEC PR #613 left these as cross-directory imports
("tactical legacy"). This commit eliminates that smell.

Symbols genuinely shared between KB-tab routes AND wizard routes
(`CookieRow`, `ConnectorSummary`) stay in -kb-types — they have a
broader scope.

Verification:
- tsc --noEmit clean (only pre-existing tsconfig warning)
- eslint . clean
