+++ klai-portal/frontend/src/routes/app/knowledge/$kbSlug_.add-connector.tsx
@@ -16,17 +16,19 @@ import * as m from '@/paraglide/messages'
 import { apiFetch } from '@/lib/apiFetch'
 import { MS_SITE_URL_PATTERN } from '@/lib/ms-docs'
 import { joinSeedUrl, ASSERTION_MODE_OPTIONS } from './$kbSlug/-kb-helpers'
-import type { CookieRow, GitHubConfig, WebCrawlerConfig } from './$kbSlug/-kb-types'
+import type { CookieRow } from './$kbSlug/-kb-types'
 import type {
   AirtableConfig,
   AuthGuardSuggestion,
   AuthProbeResult,
   ConfluenceConfig,
   ConnectorType,
+  GitHubConfig,
   NotionAddConfig,
   PreviewClassification,
   PreviewResult,
   WcStep,
+  WebCrawlerConfig,
 } from './-connector-types'
 import { MARKDOWN_PROSE_CLASSES, VALID_PRESELECT_TYPES } from './-connector-constants'
 import { AuthProbeFeedback, PreviewClassificationFeedback } from './-connector-feedback'
